pyproject.toml
[project]
name = "bench"
version = "0.1.0"
description = "A researcher's daily automation system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "feedparser>=6.0",
    "pyzotero>=1.5",
    "notion-client>=2.2",
    "google-api-python-client>=2.100",
    "google-auth-oauthlib>=1.2",
    "jinja2>=3.1",
    "rich>=13.0",
]

[project.scripts]
bench = "src.runner:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
